+++ src/storage.ts
@@ -58,6 +58,13 @@ export interface RetryOptions {
   retryableErrorFn?: (err: ApiError) => boolean;
 }
 
+export interface PreconditionOptions {
+  ifGenerationMatch?: number;
+  ifGenerationNotMatch?: number;
+  ifMetagenerationMatch?: number;
+  ifMetagenerationNotMatch?: number;
+}
+
 export interface StorageOptions extends ServiceOptions {
   retryOptions?: RetryOptions;
   autoRetry?: boolean; //Deprecated. Use retryOptions instead.
@@ -477,6 +484,10 @@ export class Storage extends Service {
    *     error should be retried and false otherwise.
    * @property {string} [userAgent] The value to be prepended to the User-Agent
    *     header in API requests.
+   * @property {object} [authClient] GoogleAuth client to reuse instead of creating a new one.
+   * @property {number} [timeout] The amount of time in milliseconds to wait per http request before timing out.
+   * @property {object[]} [interceptors_] Array of custom request interceptors to be returned in the order they were assigned.
+   * @property {string} [apiEndpoint = storage.google.com] The API endpoint of the service used to make requests.
    */
   /**
    * Constructs the Storage client.